Função 5 funcionando, precisa ajustar a formatação
