+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\MorseSimulator"
+  ],
+  "SelectedNode": "\\MorseSimulator\\morse.c",
+  "PreviewInSolutionExplorer": false
+}
