--- ConsoleApplication1/pch.cpp
@@ -1,5 +1,0 @@
-// pch.cpp: arquivo de origem correspondente ao cabeçalho pré-compilado; necessário para que a compilação seja bem-sucedida
-
-#include "pch.h"
-
-// Em geral, ignore este arquivo, mas mantenha-o se você estiver usando cabeçalhos pré-compilados.
