+++ MorseSimulator.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29306.81
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MorseSimulator", "MorseSimulator\MorseSimulator.vcxproj", "{B978BC5F-2FAE-496A-90A6-8B6218E3817C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Debug|x64.ActiveCfg = Debug|x64
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Debug|x64.Build.0 = Debug|x64
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Debug|x86.ActiveCfg = Debug|Win32
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Debug|x86.Build.0 = Debug|Win32
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Release|x64.ActiveCfg = Release|x64
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Release|x64.Build.0 = Release|x64
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Release|x86.ActiveCfg = Release|Win32
+		{B978BC5F-2FAE-496A-90A6-8B6218E3817C}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2B21EAE6-0DA5-4C22-82D0-F58CF1805ABE}
+	EndGlobalSection
+EndGlobal
