+++ MorseSimulator/Debug/vcpkg.applocal.log
@@ -1,0 +1,1 @@
+﻿
