+++ MorseSimulator/codigo.h
@@ -1,0 +1,14 @@
+#ifndef FUNCTIONS_H_INCLUDED
+#define FUNCTIONS_H_INCLUDED
+
+#define DUR 70 // DURAÇÃO DO BEEP CURTO (PONTO) EM MILISEGUNDOS
+#define TAMFRASE 9964 // TAMANHO MAXIMO DA FRASE
+
+void INTRACHAR(void);
+void ENTRECHAR(void);
+void ENTREPALAVRA(void);
+void ponto(void);
+void linha(void);
+void codigo(caractere);
+
+#endif
