+++ MorseSimulator/function5.h
@@ -1,0 +1,10 @@
+#ifndef FUNCTIONS_H_INCLUDED
+#define FUNCTIONS_H_INCLUDED
+
+#define DUR 75 // DURAÇÃO DO BEEP CURTO (PONTO) EM MILISEGUNDOS
+#define TAMFRASE 1024 // TAMANHO MAXIMO DA FRASE
+
+char* convert(char* mword);
+void parsemorse(char phrase[]);
+
+#endif
