--- codigos.h
@@ -1,18 +1,0 @@
-#ifndef FUNCTIONS_H_INCLUDED
-#define FUNCTIONS_H_INCLUDED
-
-#define DUR 75 // DURAÇÃO DO BEEP CURTO (PONTO) EM MILISEGUNDOS
-#define TAMFRASE 1024 // TAMANHO MAXIMO DA FRASE
-/* ^^ these are the include guards */
-
-/* Prototypes for the functions */
-/* Sums two ints */
-void INTRACHAR(void);
-void ENTRECHAR(void);
-void ENTREPALAVRA(void);
-void ponto(void);
-void linha(void);
-void codigo(caractere);
-
-
-#endif
